+++ src/components/Icon/icon.ts
@@ -1,5 +1,4 @@
 import { ReactComponent as BackIc } from 'assets/svgs/backBtn.svg';
-import { ReactComponent as CardIc } from 'assets/svgs/card.svg';
 import { ReactComponent as ClockIc } from 'assets/svgs/clock.svg';
 import { ReactComponent as DropDownIc } from 'assets/svgs/dropDownBtn.svg';
 import { ReactComponent as DropUpIc } from 'assets/svgs/dropUpBtn.svg';
@@ -7,15 +6,12 @@ import { ReactComponent as ExitIc } from 'assets/svgs/exitIcon.svg';
 import { ReactComponent as HambergerIc } from 'assets/svgs/hambergerIcon.svg';
 import { ReactComponent as InputCancelIc } from 'assets/svgs/inputCancelBtn.svg';
 import { ReactComponent as InputErrorIc } from 'assets/svgs/inputErrorBtn.svg';
-import { ReactComponent as InsertIc } from 'assets/svgs/insert.svg';
 import { ReactComponent as LinkIc } from 'assets/svgs/linkIcon.svg';
 import { ReactComponent as MainLogoIc } from 'assets/svgs/mainLogo.svg';
-import { ReactComponent as MakeIc } from 'assets/svgs/make.svg';
 import { ReactComponent as PasswordEyeIc } from 'assets/svgs/passwordEye.svg';
 import { ReactComponent as PasswordOpenEyeIc } from 'assets/svgs/passwordOpenEye.svg';
 import { ReactComponent as PlaceIc } from 'assets/svgs/placeIcon.svg';
 import { ReactComponent as PlusIc } from 'assets/svgs/plusBtn.svg';
-import { ReactComponent as PointIc } from 'assets/svgs/point.svg';
 import { ReactComponent as RadioCheckIc } from 'assets/svgs/radioCheck.svg';
 import { ReactComponent as RadioCheckedIc } from 'assets/svgs/radioChecked.svg';
 
@@ -36,8 +32,4 @@ export {
   LinkIc,
   PlaceIc,
   ClockIc,
-  CardIc,
-  InsertIc,
-  MakeIc,
-  PointIc
 };
