+++ src/pages/selectSchdule/components/dummyData.ts
@@ -1,0 +1,35 @@
+export const dummyData = {
+  duration: 'HOUR_HALF',
+  place: 'ONLINE',
+  placeDetail: '구글미트',
+  availableDates: [
+    {
+      month: 7,
+      day: 6,
+      dayOfWeek: '월',
+    },
+    {
+      month: 7,
+      day: 7,
+      dayOfWeek: '화',
+    },
+    {
+      month: 7,
+      day: 9,
+      dayOfWeek: '수',
+    },
+    {
+      month: 7,
+      day: 15,
+      dayOfWeek: '목',
+    },
+  ],
+  preferTimes: [
+    {
+      startTime: '06:00',
+      endTime: '12:00',
+    },
+  ],
+};
+
+export const time = ['11:00', '11:30', '12:00', '12:30', '13:00', '13:30', '14:00'];
