+++ src/pages/selectSchdule/types/Schedule.ts
@@ -1,8 +1,10 @@
-export interface DateStates {
+export interface SelectBoxState {
   id: number;
+  dropDown: boolean;
+}
+
+export interface ScheduleStates {
   date: string;
   startTime: string;
   endTime: string;
-  dropDown: boolean;
-  priority: number;
 }
