+++ src/recoil/atoms/dateAtom.ts
@@ -1,0 +1,16 @@
+import { atom } from 'recoil';
+
+export const scheduleAtoms = atom({
+  key: 'scheduleAtoms',
+  default: [
+    {
+      date: '',
+      startTime: '',
+      endTime: '',
+    },
+  ],
+});
+export const scheduleAtom = atom({
+  key: 'scheduleAtom',
+  default: { date: '', startTime: '', endTime: '' },
+});
