+++ vite.config.ts
@@ -3,10 +3,10 @@ import path from 'path';
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import macrosPlugin from 'vite-plugin-babel-macros';
-import svgr from 'vite-plugin-svgr'
+import svgr from 'vite-plugin-svgr';
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), macrosPlugin(),svgr()],
+  plugins: [react(), macrosPlugin(), svgr()],
 
   resolve: {
     alias: [
@@ -15,7 +15,7 @@ export default defineConfig({
       { find: 'components', replacement: path.resolve(__dirname, 'src/components') },
       { find: 'assets', replacement: path.resolve(__dirname, 'src/assets') },
       { find: 'styles', replacement: path.resolve(__dirname, 'src/styles') },
-      { find: 'hooks', replacement: path.resolve(__dirname, 'src/hooks') }
+      { find: 'hooks', replacement: path.resolve(__dirname, 'src/hooks') },
     ],
   },
 });
